# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MoneyApp API
  version: 1.0.0
  description: |
    Self-hosted budget and net-worth tracker.

    Conventions:
      * Session cookie auth. `user_id` is taken only from the session, never from input.
      * Money is always an object with integer minor units. No floats on the wire.
      * `null` means "not recorded". `0` means "recorded zero". Never interchangeable.
      * Errors are RFC 7807 application/problem+json.
      * Lists are cursor-paginated on (occurred_on, id).
  license:
    name: MIT
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /transactions:
    get:
      tags: [transactions]
      summary: List transactions
      parameters:
        - { name: from,        in: query, schema: { type: string, format: date } }
        - { name: to,          in: query, schema: { type: string, format: date } }
        - { name: category_id, in: query, schema: { type: integer } }
        - { name: account_id,  in: query, schema: { type: integer } }
        - name: kind
          in: query
          schema: { type: string, enum: [expense, income, transfer_in, transfer_out] }
        - { name: q,      in: query, schema: { type: string }, description: Free-text search over description and merchant }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit,  in: query, schema: { type: integer, default: 50, maximum: 500 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransactionPage" }
    post:
      tags: [transactions]
      summary: Create a transaction
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "422": { $ref: "#/components/responses/Problem" }
  /transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [transactions]
      summary: Get a transaction
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [transactions]
      summary: Update a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
    delete:
      tags: [transactions]
      summary: Soft-delete a transaction
      responses:
        "204": { description: Deleted }
  /transfers:
    post:
      tags: [transactions]
      summary: Create a transfer
      description: Creates a paired transfer_out and transfer_in sharing one transfer_group_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_account_id, to_account_id, occurred_on, amount]
              properties:
                from_account_id: { type: integer }
                to_account_id:   { type: integer }
                occurred_on:     { type: string, format: date }
                amount:          { $ref: "#/components/schemas/Money" }
                description:     { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: moneyapp_session
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string }
  responses:
    Problem:
      description: RFC 7807 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:   { type: string }
        title:  { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }
    Money:
      type: object
      required: [amount_minor, currency, exponent]
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Value in minor units. HUF has exponent 0, so 1000 means 1000 HUF.
        currency: { type: string, minLength: 3, maxLength: 3 }
        exponent: { type: integer, minimum: 0, maximum: 4 }
    BaseMoney:
      allOf:
        - $ref: "#/components/schemas/Money"
        - type: object
          properties:
            fx_rate_id: { type: integer, nullable: true }
            as_of_date: { type: string, format: date, nullable: true }
...
    Transaction:
      type: object
      properties:
        id:            { type: integer }
        account_id:    { type: integer }
        account_name:  { type: string }
        category_id:   { type: integer, nullable: true }
        category_name: { type: string }
        occurred_on: { type: string, format: date }
        kind:        { type: string, enum: [expense, income, transfer_in, transfer_out] }
        amount:      { $ref: "#/components/schemas/Money" }
        base_amount: { $ref: "#/components/schemas/BaseMoney" }
        description: { type: string, nullable: true }
        merchant:    { type: string, nullable: true }
        occurrence:  { type: integer }
        unconverted:
          type: boolean
          description: |
            True when base_amount is null because no rate covered this date. The
            write is never blocked by a missing rate, so the client is told instead.
        transfer_group_id: { type: integer, nullable: true }
        import_batch_id:   { type: integer, nullable: true }
    TransactionInput:
      type: object
      required: [account_id, occurred_on, kind, amount]
      properties:
        account_id:  { type: integer }
        category_id:
          type: integer
          nullable: true
          description: Required for expense and income; forbidden for transfers.
        occurred_on: { type: string, format: date }
        kind:        { type: string, enum: [expense, income] }
        amount:      { $ref: "#/components/schemas/Money" }
        description: { type: string }
        merchant:    { type: string }
      additionalProperties: false
    TransactionPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Transaction" }
        next_cursor: { type: string, nullable: true }
        has_more:    { type: boolean }
...